Add utility functions for event and task date formatting, and refactor related components to utilize these utilities. Introduce color coding for date differences in task and event displays, enhancing visual feedback for due dates.
