+++ lib/common/utils/date_time_utils.dart
@@ -1,9 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:zoey/core/theme/colors/app_colors.dart';
-import 'package:zoey/l10n/generated/l10n.dart';
-import 'package:zoey/features/events/models/events_model.dart';
-import 'package:zoey/features/task/models/task_model.dart';
 
 class DateTimeUtils {
   static const String dateFormat = 'd MMM yyyy';
@@ -39,82 +35,9 @@ class DateTimeUtils {
     return DateTime(date.year, date.month, date.day, time.hour, time.minute);
   }
 
-  static String getCurrentDateFormatted({String? format}) {
+  static String getTodayDateFormatted({String? format}) {
     final now = DateTime.now();
     return DateFormat(format ?? dateFormatWithDay).format(now);
-  }
-
-  static String formatEventDateAndTime(BuildContext context, EventModel event) {
-    final now = DateTime.now();
-    final today = DateTime(now.year, now.month, now.day);
-    final eventDate = DateTime(
-      event.startDate.year,
-      event.startDate.month,
-      event.startDate.day,
-    );
-
-    String dayText;
-    if (eventDate == today) {
-      dayText = L10n.of(context).today;
-    } else if (eventDate == today.add(const Duration(days: 1))) {
-      dayText = L10n.of(context).tomorrow;
-    } else {
-      dayText = DateFormat('EEE').format(eventDate);
-    }
-
-    final timeText = DateFormat(timeFormat).format(event.startDate);
-    return L10n.of(context).dayAtTime(dayText, timeText);
-  }
-
-  static String formatTaskDueDate(BuildContext context, TaskModel task) {
-    final now = DateTime.now();
-    final today = DateTime(now.year, now.month, now.day);
-    final dueDate = DateTime(
-      task.dueDate.year,
-      task.dueDate.month,
-      task.dueDate.day,
-    );
-    final difference = dueDate.difference(today).inDays;
-
-    if (difference < 0) {
-      final overdueDays = -difference;
-      return L10n.of(context).overdueByDays(
-        overdueDays == 1
-            ? L10n.of(context).oneDay
-            : L10n.of(context).daysCounts(overdueDays),
-      );
-    } else if (difference == 0) {
-      return L10n.of(context).dueToday;
-    } else if (difference == 1) {
-      return L10n.of(context).dueTomorrow;
-    } else if (difference <= 7) {
-      return L10n.of(context).dueInDays(difference);
-    } else {
-      return L10n.of(context).dueDayMonth(task.dueDate.day, task.dueDate.month);
-    }
-  }
-
-  static Color getTaskDueDateColor(TaskModel task) {
-    final now = DateTime.now();
-    final today = DateTime(now.year, now.month, now.day);
-    final dueDate = DateTime(
-      task.dueDate.year,
-      task.dueDate.month,
-      task.dueDate.day,
-    );
-    final difference = dueDate.difference(today).inDays;
-
-    if (difference < 0) {
-      return AppColors.errorColor;
-    } else if (difference == 0) {
-      return AppColors.brightOrangeColor;
-    } else if (difference == 1) {
-      return AppColors.brightMagentaColor;
-    } else if (difference <= 3) {
-      return AppColors.primaryColor;
-    } else {
-      return AppColors.successColor;
-    }
   }
 
   static Future<DateTime?> showDatePickerDialog(
