+++ lib/features/events/widgets/event_widget.dart
@@ -1,26 +1,25 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:zoey/common/utils/date_time_utils.dart';
+import 'package:zoey/common/utils/common_utils.dart';
 import 'package:zoey/common/widgets/toolkit/zoe_close_button_widget.dart';
 import 'package:zoey/common/widgets/toolkit/zoe_inline_text_edit_widget.dart';
 import 'package:zoey/core/routing/app_routes.dart';
 import 'package:zoey/features/events/models/events_model.dart';
 import 'package:zoey/features/events/providers/events_proivder.dart';
+import 'package:zoey/features/events/utils/event_utils.dart';
 import 'package:zoey/features/events/widgets/event_date_widget.dart';
 import 'package:zoey/l10n/generated/l10n.dart';
 
 class EventWidget extends ConsumerWidget {
   final String eventsId;
   final bool isEditing;
-  final bool isTodayFocusView;
   final EdgeInsetsGeometry? margin;
 
   const EventWidget({
     super.key,
     required this.eventsId,
     required this.isEditing,
-    this.isTodayFocusView = false,
     this.margin,
   });
 
@@ -103,23 +102,12 @@ class EventWidget extends ConsumerWidget {
     WidgetRef ref,
     EventModel event,
   ) {
-    final theme = Theme.of(context);
-    if (isTodayFocusView) {
-      return Text(
-        DateTimeUtils.formatEventDateAndTime(context, event),
-        style: theme.textTheme.bodySmall?.copyWith(
-          color: theme.colorScheme.secondary,
-          fontWeight: FontWeight.w500,
-        ),
-      );
-    }
-    final startDateText = DateTimeUtils.formatDateTime(event.startDate);
-    final endDateText = DateTimeUtils.formatDateTime(event.endDate);
     return Text(
-      '$startDateText - $endDateText',
-      maxLines: 2,
-      overflow: TextOverflow.ellipsis,
-      style: Theme.of(context).textTheme.bodySmall,
+      EventUtils.formatEventDateAndTime(context, event),
+      style: Theme.of(context).textTheme.bodySmall?.copyWith(
+        color: CommonUtils.getColorByDateDifference(event.startDate),
+        fontWeight: FontWeight.w500,
+      ),
     );
   }
